backend/requirements.txt
# ============================================================================
# File Path: backend/requirements.txt
# Description: Python dependencies for Django backend
# ============================================================================

Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
boto3==1.34.34
Pillow==10.2.0
pytest==8.0.1
pytest-django==4.8.0
pytest-cov==4.1.0
flake8==7.0.0
black==24.2.0
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
python-dotenv==1.0.1 